webpack 1

+++ controllers/videoController.js
@@ -25,21 +25,31 @@ export const watch = async (req,res) => {
 
 export const getEdit = async (req,res) => {
     const {id} = req.params
+    const {user: {_id}} = req.session
     const video = await Video.findById(id)
 
     if(!video) {
         return res.status(404).render("404", {pageTitle: "Video not found."})
     }
+    if(String(video.owner) !== String(_id)) {
+        return res.staus(403).redirect('/')
+    }
+    
     return res.render("edit", {pageTitle: `Editing`, video})
 }
 
 export const postEdit = async (req,res) => {
+    const {user: {_id}} = req.session
     const {id} = req.params
     const {title, description, hashtags} = req.body
     const video = await Video.exists({_id: id})
 
     if(!video) {
         return res.status(404).render("404", {pageTitle: "Video not found."})
+    }
+
+    if(String(video.owner) !== String(_id)) {
+        return res.staus(403).redirect('/')
     }
 
     await Video.findByIdAndUpdate(id, {
@@ -59,8 +69,7 @@ export const postUpload = async(req,res) => {
     const {path: fileUrl} = req.file
     const {title, description, hashtags} = req.body
     try {
-        console.log(req.body); // get the data   from post request
-        await Video.create({
+        const newVideo = await Video.create({
             title, // same with title:title(req.body.title)
             description,
             fileUrl,
@@ -70,11 +79,11 @@ export const postUpload = async(req,res) => {
             // : hashtags.split(",").map((word) => `#${word}`)
             // ,를 기준으로 array 안의 data를 split을 통해 독립적으로 분리시키고
             // map을 통해 각각의 분리된 단어들에 #을 추가한다.
-            meta: {
-                views: 0,
-                rating: 0,
-            }
         })
+        const userId = await user.findById(_id)
+        userId.videos.push(newVideo._id)
+        userId.save()
+
         return res.redirect("/")
     } catch (error) {
         return res.status(400).render("upload", {
@@ -86,6 +95,14 @@ export const postUpload = async(req,res) => {
 
 export const deleteVideo = async (req,res) => {
     const {id} = req.params
+    const {user: {_id}} = req.session
+    const video = await Video.findById(id)
+    if(!video) {
+        return res.status(404).render("404", {pageTitle: "Video not found."})
+    }
+    if(String(video.owner) !== String(_id)) {
+        return res.staus(403).redirect('/')
+    }
     await Video.findByIdAndDelete(id)
     // delete video
     return res.redirect("/")
